+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * `is_base_of` no longer caught up on access modifierrs
 * General c++03 compatibility
 * `estd::optional` and `estd::expected` much more robust, shares code base with `estd::variant`
+* `estd::num_get` and by extension `istream` integer parsing now returns min/max value on overflow
 
 # v0.5.1 - 18MAY23